Docker build speedup
Re-order docker file so that layers which are unlikely to change between
builds appear first. Layers are now: clamav packages, pip requirements,
and finally lambda code. This means that changing lambda code will not
require re-installing clamav packages or rebuilding the rest of the
image.

Added a cleandocker target to the Makefile in order to explicitly
rebuild from scratch (useful for when an update to clamav packages is
required)